requirements.txt
psycopg2-binary
python-dotenv
SQLAlchemy
web3
pyTelegramBotAPI
